disable edit send if policy requires
