+++ libs/tools/send/send-ui/src/send-form/components/options/send-options.component.ts
@@ -1,8 +1,8 @@
 import { CommonModule } from "@angular/common";
-import { Component, Input, OnInit } from "@angular/core";
+import { Component, Input, OnDestroy, OnInit } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 import { FormBuilder, ReactiveFormsModule } from "@angular/forms";
-import { map } from "rxjs";
+import { map, Subject, takeUntil } from "rxjs";
 
 import { JslibModule } from "@bitwarden/angular/jslib.module";
 import { PolicyService } from "@bitwarden/common/admin-console/abstractions/policy/policy.service.abstraction";
@@ -38,7 +38,9 @@ import { SendFormContainer } from "../../send-form-container";
     CommonModule,
   ],
 })
-export class SendOptionsComponent implements OnInit {
+export class SendOptionsComponent implements OnInit, OnDestroy {
+  private destroy$ = new Subject<void>();
+
   @Input({ required: true })
   config: SendFormConfig;
   @Input()
@@ -97,6 +99,7 @@ export class SendOptionsComponent implements OnInit {
       });
     });
   }
+
   ngOnInit() {
     if (this.sendFormContainer.originalSendView) {
       this.sendOptionsForm.patchValue({
@@ -107,5 +110,18 @@ export class SendOptionsComponent implements OnInit {
         notes: this.sendFormContainer.originalSendView.notes,
       });
     }
+    this.policyService
+      .policyAppliesToActiveUser$(PolicyType.DisableSend)
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((policyAppliesToActiveUser) => {
+        if (policyAppliesToActiveUser) {
+          this.sendOptionsForm.disable();
+        }
+      });
+  }
+
+  ngOnDestroy() {
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 }
