+++ libs/tools/send/send-ui/src/send-form/components/send-form.component.ts
@@ -182,6 +182,8 @@ export class SendFormComponent implements AfterViewInit, OnInit, OnChanges, Send
     this.loading = false;
   }
 
+  formDisabled = false;
+
   constructor(
     private formBuilder: FormBuilder,
     private addEditFormService: SendFormService,
